Fix verbose request printer
Since the body was read during the printing (if the verbose flag was
set), the io.Reader was drained when the request was sent (despite the
correct Content-Length), causing a client.Do() error.  This was fixed using the
httputil.Dump(Request|Response) functions.  Though TeeReader could help
too in the future, these changes satisfy my needs.