Refactoring the utils packages back into main.
The cliutils and httputils were not really 'packages' in the golang
sense in that they were not standalone libraries and depended on the
logic being developed for the gocurl http client.  So instead of
separated packages, they are individual files in the main package.
This will actually allow for some clean up.